+++ src/main/java/fr/maxlego08/essentials/scoreboard/ZEssentialsScoreboard.java
@@ -4,6 +4,7 @@ import fr.maxlego08.essentials.api.scoreboard.EssentialsScoreboard;
 import fr.maxlego08.essentials.api.scoreboard.PlayerBoard;
 import fr.maxlego08.essentials.api.scoreboard.ScoreboardAnimationType;
 import fr.maxlego08.essentials.api.scoreboard.ScoreboardLine;
+import fr.maxlego08.essentials.api.scoreboard.ScoreboardType;
 import fr.maxlego08.essentials.api.scoreboard.configurations.ColorWaveConfiguration;
 import fr.maxlego08.essentials.api.scoreboard.configurations.NoneConfiguration;
 import fr.maxlego08.essentials.zutils.utils.ZUtils;
@@ -12,25 +13,32 @@ import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Player;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 public class ZEssentialsScoreboard extends ZUtils implements EssentialsScoreboard {
+
     private final String name;
     private final boolean isDefault;
     private final String title;
     private final List<ScoreboardLine> lines = new ArrayList<>();
+    private final ScoreboardType scoreboardType;
 
-    public ZEssentialsScoreboard(String name, boolean isDefault, String title, List<ScoreboardLine> lines) {
+    public ZEssentialsScoreboard(String name, boolean isDefault, String title, List<ScoreboardLine> lines, ScoreboardType scoreboardType) {
         this.name = name;
         this.isDefault = isDefault;
         this.title = title;
+        this.scoreboardType = scoreboardType;
     }
 
     public ZEssentialsScoreboard(String scoreboardName, ConfigurationSection configurationSection) {
         this.name = scoreboardName;
         this.isDefault = configurationSection.getBoolean("default", false);
         this.title = configurationSection.getString("title", "");
+        this.scoreboardType = ScoreboardType.valueOf(configurationSection.getString("type", ScoreboardType.NORMAL.name()).toUpperCase());
+
         configurationSection.getMapList("lines").forEach(currentLine -> {
 
             TypedMapAccessor accessor = new TypedMapAccessor((Map<String, Object>) currentLine);
@@ -84,13 +92,37 @@ public class ZEssentialsScoreboard extends ZUtils implements EssentialsScoreboar
 
     @Override
     public void create(PlayerBoard playerBoard) {
+        this.update(playerBoard);
+
+        this.lines.forEach(line -> line.startAnimation(playerBoard));
+    }
+
+    @Override
+    public void update(PlayerBoard playerBoard) {
+
         Player player = playerBoard.getPlayer();
 
         playerBoard.updateTitle(papi(this.title, player));
 
-        List<String> lines = this.lines.stream().map(ScoreboardLine::getText).map(line -> papi(line, player)).toList();
-        playerBoard.updateLines(lines);
+        List<String> lines = new ArrayList<>();
+        int line = 0;
+        for (ScoreboardLine scoreboardLine : this.lines) {
+            String[] split = scoreboardLine.getText().split("\n");
+            for (String s : split) {
+                lines.add(line, papi(s, playerBoard.getPlayer()));
+                line++;
+            }
+            scoreboardLine.setLine(line - split.length);
+        }
 
-        this.lines.forEach(line -> line.startAnimation(playerBoard));
+        playerBoard.updateLines(lines);
+    }
+
+    @Override
+    public void update(PlayerBoard playerBoard, String eventName) {
+
+        if (this.scoreboardType != ScoreboardType.NORMAL) return;
+
+        this.lines.stream().filter(scoreboardLine -> scoreboardLine.getEventName() != null && scoreboardLine.getEventName().equals(eventName)).forEach(scoreboardLine -> scoreboardLine.update(playerBoard));
     }
 }